Hack up sub-bottom qualifier sets.
These are used for nullary type constructors, which are the only case where we push inconsistent facts to the theorem prover continuously.  This causes troubles when a nullary type constructor's value can reach the guard of an if, because we'll push this fact as we examine the body of the if, causing unsoundness.