Tighten up the var rule a bit; move working examples from broken/ to small/
The new var rule basically says that, if the var is an identifier, we can use the type {a | a = x} and let the resulting type of the expression be a supertype of that.  Since the var must be in scope, it's in the environment, so anything implied by its type will be implied by this type as well.  I admit this is very strange.