Fix application of non-var-patterned functions; bugfixes; make Pattern module
Support functions with non-variable (e.g., maybe i.e., tuple) patterns.

When a function of the form (a1, ..., an) -> P(a1, ..., an) is applied, we need
to take care to substitute all vars in the tuple.  Otherwise, if we're in a
context where some ai exists, we risk being unsound.

bind_pat had a bug in the Tpat_any case.

Some useful functions are now in the Pattern module.